Show testmerge changelogs, convert to TypeScript
